+++ libasc/src/hash.rs
@@ -1,12 +1,14 @@
 use std::{fmt::{Debug, Display, Formatter}, str::FromStr};
 
 use eyre::bail;
+use rateless_tables::{Bytes, Symbol};
 use serde::{Deserialize, Serialize};
 
 pub type RawObjectHash = [u8; 32];
 
 /// A SHA-256 wrapper type used to uniquely identify content in the repository.
-#[derive(Clone, Copy, Default, Deserialize, Eq, Hash, PartialEq, PartialOrd, Ord, Serialize)]
+#[derive(Clone, Copy, Default, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
+#[repr(transparent)]
 pub struct ObjectHash(#[serde(with = "serde_bytes")] RawObjectHash);
 
 impl ObjectHash {
@@ -30,7 +32,7 @@ impl ObjectHash {
     }
 
     /// Get the individual bytes that make up this `ObjectHash`.
-    pub fn as_bytes(&self) -> &[u8] {
+    pub fn as_bytes(&self) -> &RawObjectHash {
         &self.0
     }
 }
@@ -64,5 +66,29 @@ impl FromStr for ObjectHash {
         }
 
         Ok(ObjectHash(bytes[..].try_into()?))
+    }
+} 
+
+impl Bytes for ObjectHash {
+    fn from_bytes(bytes: &[u8]) -> Self {
+        let bytes: RawObjectHash = bytes.try_into().expect("could not create hash from bytes");
+
+        bytes.into()
+    }
+
+    fn to_bytes(&self) -> Vec<u8> {
+        self.as_bytes().to_vec()
+    }
+}
+
+impl Symbol for ObjectHash {
+    fn xor(&self, other: &Self) -> Self {
+        let mut result = *self.as_bytes();
+
+        for (i, &v) in other.as_bytes().iter().enumerate() {
+            result[i] ^= v;
+        }
+
+        result.into()
     }
 }