+++ libasc/src/key.rs
@@ -1,84 +1,17 @@
-use std::{fmt::{Debug, Display, Formatter, Result as FmtResult}, ops::{Deref, DerefMut}};
+use std::{fmt::{Debug, Display, Formatter, Result as FmtResult}, hash::Hash, ops::{Deref, DerefMut}};
 
 use crate::unwrap;
 
-use ecdsa::{SigningKey, VerifyingKey, signature::{SignerMut, Verifier, rand_core::CryptoRngCore}};
+use ecdsa::{SigningKey, VerifyingKey, signature::{SignerMut, Verifier}};
 use eyre::Result;
-use p256::{NistP256, elliptic_curve::{NonZeroScalar, PublicKey as InnerVerifyingKey, ScalarPrimitive}};
+use p256::{NistP256};
 use rand::{CryptoRng, RngCore};
-use serde::{Deserialize, Serialize, de};
+use serde::{Deserialize, Serialize};
 use serde_bytes::ByteBuf;
 
-/// The raw version of [`PrivateKey`].
-/// 
-/// The layout of this struct is identical to that of [`SigningKey`],
-/// and is used to add custom [`serde`] logic.
-#[derive(Clone, Copy)]
-pub struct RawPrivateKey {
-    _secret_scalar: NonZeroScalar<NistP256>,
-    verifying_key: VerifyingKey<NistP256>
-}
-
-impl RawPrivateKey {
-    /// Generate a random [`RawPrivateKey`] with the given RNG device.
-    pub fn random(rng: &mut impl CryptoRngCore) -> RawPrivateKey {
-        SigningKey::<NistP256>::random(rng).into()
-    }
-
-    /// Reconstruct a [`RawPrivateKey`] from raw bytes.
-    pub fn from_bytes(bytes: &[u8]) -> Result<RawPrivateKey> {
-        let raw = SigningKey::<NistP256>::from_bytes(bytes.into())?;
-        
-        Ok(raw.into())
-    }
-}
-
-impl From<SigningKey<NistP256>> for RawPrivateKey {
-    fn from(value: SigningKey<NistP256>) -> Self {
-        unsafe { std::mem::transmute(value) }
-    }
-}
-
-impl Debug for RawPrivateKey {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        write!(f, "RawPrivateKey {{ inner: SigningKey {{ secret_scalar: ..., verifying_key: {:?} }} }}", self.verifying_key)
-    }
-} 
-
-impl Deref for RawPrivateKey {
-    type Target = SigningKey<NistP256>;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl Serialize for RawPrivateKey {
-    fn serialize<S: serde::Serializer>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error> {
-        let bytes: &[u8] = &self.to_bytes();
-
-        let buf = ByteBuf::from(bytes);
-
-        buf.serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for RawPrivateKey {
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> std::result::Result<Self, D::Error> {
-        let bytes = ByteBuf::deserialize(deserializer)?;
-
-        let priv_key = PrivateKey::from_bytes(&bytes)
-            .map_err(|_| de::Error::invalid_value(de::Unexpected::Bytes(&bytes), &"a valid private key"))?;
-
-        let raw: RawPrivateKey = priv_key.0;
-
-        Ok(raw)
-    }
-}
-
 /// A private key used for creating signatures.
-#[derive(Clone, Copy, Debug, Deserialize, Serialize)]
-pub struct PrivateKey(RawPrivateKey);
+#[derive(Clone, Debug)]
+pub struct PrivateKey(SigningKey<NistP256>);
 
 impl PrivateKey {
     /// Create a new random [`PrivateKey`].
@@ -91,26 +24,20 @@ impl PrivateKey {
 
     /// Create a [`PrivateKey`] from a given RNG device.
     pub fn random<C: CryptoRng + RngCore>(rng: &mut C) -> PrivateKey {
-        PrivateKey(RawPrivateKey::random(rng))
+        PrivateKey(SigningKey::<NistP256>::random(rng))
     }
 
     /// Get the corresponding [`PublicKey`] to this [`PrivateKey`].
     pub fn public_key(&self) -> PublicKey {
-        let raw: RawPublicKey = (*self.0.verifying_key()).into();
+        let raw: VerifyingKey<NistP256> = *self.0.verifying_key();
 
         raw.into()
     }
 
     /// Sign some data using the private key.
     pub fn sign(&mut self, bytes: &[u8]) -> Signature {
-        let raw_key: &mut RawPrivateKey = &mut *self;
-        
-        let lib_sig: ecdsa::Signature<NistP256> = raw_key.sign(bytes);
-
-        let raw_sig: RawSignature = lib_sig.into();
-
         Signature {
-            inner: raw_sig,
+            inner: self.0.sign(bytes),
             key: self.public_key()
         }
     }
@@ -121,7 +48,7 @@ impl PrivateKey {
             "invalid length of bytes: {} (expected 32)", bytes.len()
         );
 
-        let inner = RawPrivateKey::from_bytes(&array)?;
+        let inner = SigningKey::<NistP256>::from_bytes(&array.into())?;
 
         Ok(PrivateKey(inner))
     }
@@ -132,28 +59,10 @@ impl PrivateKey {
 }
 
 impl Deref for PrivateKey {
-    type Target = RawPrivateKey;
+    type Target = SigningKey::<NistP256>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
-    }
-}
-
-impl DerefMut for RawPrivateKey {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl DerefMut for PrivateKey {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl From<RawPrivateKey> for PrivateKey {
-    fn from(value: RawPrivateKey) -> Self {
-        PrivateKey(value)
     }
 }
 
@@ -169,77 +78,36 @@ impl Display for PrivateKey {
     }
 }
 
-/// The raw version of [`PublicKey`].
-/// 
-/// The layout of this struct is identical to that of [`VerifyingKey`],
-/// and is used to add custom [`serde`] logic.
-#[derive(Clone, Copy, Debug)]
-pub struct RawPublicKey {
-    _inner: InnerVerifyingKey<NistP256>
-}
-
-impl RawPublicKey {
-    pub fn from_bytes(bytes: &[u8]) -> Result<RawPublicKey> {
-        let raw = VerifyingKey::<NistP256>::from_sec1_bytes(bytes)?;
-        
-        Ok(raw.into())
-    }
-
-    pub fn to_bytes(&self) -> Vec<u8> {
-        self.to_sec1_bytes().into_vec()
+impl Hash for PrivateKey {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        state.write(&self.to_bytes());
     }
 }
 
-impl Deref for RawPublicKey {
-    type Target = VerifyingKey<NistP256>;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl DerefMut for RawPublicKey {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl From<VerifyingKey<NistP256>> for RawPublicKey {
-    fn from(value: VerifyingKey<NistP256>) -> Self {
-        unsafe { std::mem::transmute(value) }
-    }
-}
-
-impl Serialize for RawPublicKey {
+impl Serialize for PrivateKey {
     fn serialize<S: serde::Serializer>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error> {
-        let bytes: &[u8] = &self.to_bytes();
-
-        let buf = ByteBuf::from(bytes);
+        let buf = ByteBuf::from(self.to_bytes());
 
         buf.serialize(serializer)
     }
 }
 
-impl<'de> Deserialize<'de> for RawPublicKey {
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> std::result::Result<Self, D::Error> {
-        let bytes = ByteBuf::deserialize(deserializer)?;
-
-        let pub_key = PublicKey::from_bytes(&bytes)
-            .map_err(|_| de::Error::invalid_value(de::Unexpected::Bytes(&bytes), &"a valid private key"))?;
-
-        let raw: RawPublicKey = pub_key.0;
-
-        Ok(raw)
+impl<'de> Deserialize<'de> for PrivateKey {
+    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> std::result::Result<PrivateKey, D::Error> {
+        let buf = ByteBuf::deserialize(deserializer)?;
+        
+        PrivateKey::from_bytes(buf.as_slice())
+            .map_err(serde::de::Error::custom)
     }
 }
 
 /// A public key used for verifying signatures.
-#[derive(Clone, Copy, Debug, Deserialize, Serialize)]
-pub struct PublicKey(RawPublicKey);
+#[derive(Clone, Copy, Debug, Deserialize, Eq, Serialize)]
+pub struct PublicKey(VerifyingKey<NistP256>);
 
 impl PublicKey {
     pub fn from_bytes(bytes: &[u8]) -> Result<PublicKey> {
-        let raw = RawPublicKey::from_bytes(bytes)?;
+        let raw = VerifyingKey::<NistP256>::from_sec1_bytes(bytes)?;
 
         Ok(raw.into())
     }
@@ -249,14 +117,14 @@ impl PublicKey {
     }
 }
 
-impl From<RawPublicKey> for PublicKey {
-    fn from(value: RawPublicKey) -> Self {
+impl From<VerifyingKey<NistP256>> for PublicKey {
+    fn from(value: VerifyingKey<NistP256>) -> Self {
         PublicKey(value)
     }
 }
 
 impl Deref for PublicKey {
-    type Target = RawPublicKey;
+    type Target = VerifyingKey<NistP256>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -281,67 +149,13 @@ impl Display for PublicKey {
     }
 }
 
-/// A bit-identical copy of [`VerifyingKey`].
-#[derive(Clone, Debug, PartialEq)]
-pub struct RawSignature {
-    _r: ScalarPrimitive<NistP256>,
-    _s: ScalarPrimitive<NistP256>
-}
-
-impl RawSignature {
-    pub fn from_bytes(bytes: &[u8]) -> Result<RawSignature> {
-        let inner = ecdsa::Signature::<NistP256>::from_bytes(bytes.into())?;
-
-        let signature = unsafe { std::mem::transmute::<ecdsa::Signature<NistP256>, RawSignature>(inner) };
-
-        Ok(signature)
-    }
-
-    pub fn to_bytes(&self) -> Vec<u8> {
-        self.to_vec()
+impl Hash for PublicKey {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        state.write(&self.to_bytes());
     }
 }
 
-impl Deref for RawSignature {
-    type Target = ecdsa::Signature<NistP256>;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl DerefMut for RawSignature {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { std::mem::transmute(self) }
-    }
-}
-
-impl From<ecdsa::Signature<NistP256>> for RawSignature {
-    fn from(value: ecdsa::Signature<NistP256>) -> Self {
-        unsafe { std::mem::transmute(value) }
-    }
-}
-
-impl Serialize for RawSignature {
-    fn serialize<S: serde::Serializer>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error> {
-        let bytes: &[u8] = &self.to_bytes();
-
-        let buf = ByteBuf::from(bytes);
-
-        buf.serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for RawSignature {
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> std::result::Result<Self, D::Error> {
-        let bytes = ByteBuf::deserialize(deserializer)?;
-
-        let signature = RawSignature::from_bytes(&bytes)
-            .map_err(|_| de::Error::invalid_value(de::Unexpected::Bytes(&bytes), &"a valid signature"))?;
-
-        Ok(signature)
-    }
-}
+type RawSignature = ecdsa::Signature<NistP256>;
 
 /// A signature with an attached verifying key.
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
@@ -353,7 +167,7 @@ pub struct Signature {
 impl Signature {
     /// Check if the signature is valid, returning any errors if not.
     pub fn check(&self, data: &[u8]) -> Result<()> {
-        (**self.key).verify(data, &*self.inner)?;
+        self.key.verify(data, &self.inner)?;
 
         Ok(())
     }